--- src/main/java/persistence/sql/ddl/dialect/H2Dialect.java
@@ -1,42 +1,0 @@
-package persistence.sql.ddl.dialect;
-
-import jakarta.persistence.GenerationType;
-import persistence.sql.ddl.KeyType;
-
-public class H2Dialect implements Dialect {
-    @Override
-    public String mapDataType(Class<?> type) {
-        switch (type.getSimpleName()) {
-            case "Long":
-                return "BIGINT";
-            case "String":
-                return "VARCHAR";
-            case "Integer":
-                return "INTEGER";
-            default:
-                throw new IllegalArgumentException("지원하지 않는 타입입니다.");
-        }
-    }
-
-    @Override
-    public String mapGenerationType(GenerationType strategy) {
-        switch (strategy) {
-            case IDENTITY:
-                return "generated by default as identity";
-            default:
-                throw new IllegalArgumentException("지원하지 않는 전략입니다.");
-        }
-    }
-
-    @Override
-    public String mapKeyType(KeyType keyType) {
-        switch (keyType) {
-            case PRIMARY:
-                return "PRIMARY";
-            case UNIQUE:
-                return "UNIQUE";
-            default:
-                throw new IllegalArgumentException("지원하지 않는 키 타입입니다.");
-        }
-    }
-}
